This is a test to see how commit messages work.
I am making this commmit to see how git commit messages work in VS Code.
I was tasked to do this for the Odin Project, but it was not clear how
to make multi line commit messages in the lesson. Thankfully I figured
it out. But it did kind of suck.